+++ packages/keystatic/DocumentEditor/component-blocks/utils.ts
@@ -167,6 +167,7 @@ export function getDocumentFeaturesForChildField(
       },
       links: options.links === 'inherit',
       images: options.images === 'inherit',
+      tables: options.tables === 'inherit',
     },
   };
 }