+++ packages/keystatic/markdoc/from-markdoc.ts
@@ -207,7 +207,48 @@ function fromMarkdocNode(
   if (node.type === 'hr') {
     return { type: 'divider', children: [{ text: '' }] };
   }
+  if (node.type === 'table') {
+    return {
+      type: 'table',
+      children: node.children.flatMap(x => fromMarkdocNode(x, componentBlocks)),
+    };
+  }
+  if (node.type === 'tbody') {
+    return {
+      type: 'table-body',
+      children: node.children.flatMap(x => fromMarkdocNode(x, componentBlocks)),
+    };
+  }
+  if (node.type === 'thead') {
+    if (!node.children.length) return [];
+    return {
+      type: 'table-head',
+      children: node.children.flatMap(x => fromMarkdocNode(x, componentBlocks)),
+    };
+  }
+  if (node.type === 'tr') {
+    return {
+      type: 'table-row',
+      children: node.children.flatMap(x => fromMarkdocNode(x, componentBlocks)),
+    };
+  }
+  if (node.type === 'td') {
+    return {
+      type: 'table-cell',
+      children: node.children.flatMap(x => fromMarkdocNode(x, componentBlocks)),
+    };
+  }
+  if (node.type === 'th') {
+    return {
+      type: 'table-cell',
+      header: true,
+      children: node.children.flatMap(x => fromMarkdocNode(x, componentBlocks)),
+    };
+  }
   if (node.type === 'tag') {
+    if (node.tag === 'table') {
+      return fromMarkdocNode(node.children[0], componentBlocks);
+    }
     if (node.tag === 'layout') {
       return {
         type: 'layout',