+++ packages/keystatic/markdoc/to-markdoc.tsx
@@ -223,6 +223,35 @@ function toMarkdoc(
   if (node.type === 'divider') {
     return new Ast.Node('hr');
   }
+  if (node.type === 'table') {
+    const head = node.children.find(
+      (x): x is typeof x & { type: 'table-head' } => x.type === 'table-head'
+    );
+    return new Ast.Node(
+      'tag',
+      {},
+      [
+        new Ast.Node('table', {}, [
+          new Ast.Node('thead', {}, head ? head.children.map(_toMarkdoc) : []),
+          _toMarkdoc(node.children.find(x => x.type === 'table-body')!),
+        ]),
+      ],
+      'table'
+    );
+  }
+  if (node.type === 'table-body') {
+    return new Ast.Node('tbody', {}, node.children.map(_toMarkdoc));
+  }
+  if (node.type === 'table-row') {
+    return new Ast.Node('tr', {}, node.children.map(_toMarkdoc));
+  }
+  if (node.type === 'table-cell') {
+    return new Ast.Node(
+      node.header ? 'th' : 'td',
+      {},
+      node.children.map(_toMarkdoc)
+    );
+  }
   if (node.type === 'heading') {
     const markdocNode = new Ast.Node(
       'heading',